% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NDepositsPmf.R
\name{getNDepositPmf}
\alias{getNDepositPmf}
\title{Get the pmf for number of undiscovered deposits}
\usage{
getNDepositPmf(object)
}
\arguments{
\item{object}{An object of class "NDepositsPmf"}
}
\value{
Dataframe containing the pmf. The dataframe has two columns: The
first lists the numbers of deposits. The second lists the probabilities
associated with those numbers.
}
\description{
Get the probability mass function (pmf)
for the number of undiscovered deposits in the permissive tract.
}
\examples{
nDepositsPmf <- NDepositsPmf( "Poisson", list(theMean=5), "" )
getNDepositPmf( nDepositsPmf )

}

