% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TonnagePdf1.R
\name{getRandomSamples}
\alias{getRandomSamples}
\title{Get random samples from the pdf for material tonnages in a
single, undiscovered deposit}
\usage{
getRandomSamples(object, nSamples, seed = NULL, log_rs = FALSE)
}
\arguments{
\item{object}{An object of class "TonnagePdf1"}

\item{nSamples}{Integer specifying the number of random samples}

\item{seed}{Integer containing the seed.}

\item{log_rs}{Logical variable indicating whether the random sample are transformed
with the natural logarithm.}
}
\value{
Matrix containing random samples from the pdf for the
material tonnages for a single, undiscovered deposit in the permissive
tract.
}
\description{
Get random samples from the probability density function (pdf)
for the
material tonnages in a single, undiscovered deposit within the permissive
tract.
}
\details{
To generate random samples, the observed material tonnages are
transformed with the natural logarithm.
If the \code{type} is \code{empirical}, then
a multivariate kernel density estimate
of the log-transformed tonnages is used to
generate log-transformed random samples
(Duong, 2007; Hastie and others, 2009, p. 208-209;
Shalizi, 2016, p. 308-330).
In contrast, if the \code{type} is \code{normal},
then a multivariate normal distribution is used to generate
log-transformed random samples.

The log-transformed random samples are converted to random samples of
material tonnage with exponentiation.
}
\examples{
exData <- ExampleTonnageData
pdf1 <- TonnagePdf1(exData, "mt")
rs <- getRandomSamples(pdf1, 2518)

}
\references{
Duong, Tarn, 2007, ks - Kernel density estimation and kernel discriminant
analysis for multivariate data in R: Journal of Statistical Software,
v. 21, issue 7, \url{http://www.jstatsoft.org/}

Hastie, Tevor, Tibshirani, Robert, and Friedman, Jerome, 2009,
The elements of statistical learning - Data mining, inference, and
prediction, 2nd ed., Springer Science + Business Media, LLC, 745 p.

Shalizi, C.R., 2016, Advanced data analysis from an elementary point of
view: Draft book manuscript publicly available at
\url{http://www.stat.cmu.edu/~cshalizi/ADAfaEPoV/}
}

