% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Metadata.R
\name{getTractId}
\alias{getTractId}
\title{Get the tract identifier}
\usage{
getTractId(object)
}
\arguments{
\item{object}{An object of class "Metadata"}
}
\value{
A character string containing the tract identifier.
}
\description{
Get the tract identifier, which is a unique character string
specifying the permissive tract.
}
\examples{
meta <- Metadata("Lucky Strike PT", "PT001")
cat(sprintf("Tract identifier: \%s\\n", getTractId(meta)))

}

