% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TonnagePdf1.R
\name{getUnits}
\alias{getUnits}
\title{Get the units for tonnage}
\usage{
getUnits(object)
}
\arguments{
\item{object}{An object of class "TonnagePdf1"}
}
\value{
A character string containing the units.
}
\description{
Get the units for tonnage.
}
\examples{
exData <- ExampleTonnageData
pdf1 <- TonnagePdf1(exData, "mt")
cat(sprintf("Units for tonnage: \%s\\n", getUnits(pdf1)))

}

