% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Metadata.R
\name{Metadata}
\alias{Metadata}
\title{Construct the metadata for the probability calculations}
\usage{
Metadata(tractName, tractId, tractArea = "", nKnownDeposits = 0,
  description = "")
}
\arguments{
\item{tractName}{Character string containing the name of the permissive tract.}

\item{tractId}{Character string containing the tract identifier, which is a unique
identifier for the permissive tract. (See Details).}

\item{tractArea}{Character string containing the area of the permissive tract and the units
for the area.}

\item{nKnownDeposits}{Integer containing the number of known deposits in the permissive tract.}

\item{description}{A description of the permissive tract.}
}
\value{
A list with the following components is returned.

\item{tractName}{Input argument tractName}

\item{tractId}{Input argument tractId}

\item{tractArea}{Input argument tractArea}

\item{nKnownDeposits}{Input argument nKnownDeposits}

\item{description}{Input argument description}

\item{call}{Function call}
}
\description{
Construct the metadata for the probability calculations
for one permissive tract.
}
\details{
Parameter \code{tractId}, which is a character string, may be used to file
names associated with the probability calculations. Consequently, it is
recommended that \code{tractId} be suitable for a file name. That is,
\code{tractId}
should contain neither spaces nor odd characters such as !, $, and &.
}
\examples{
meta <- Metadata("Lucky Strike PT", "PT001", "mt")

}

