% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NDepositsPmf.R
\name{plot.NDepositsPmf}
\alias{plot.NDepositsPmf}
\title{Plot the pmf for the number of undiscovered deposits}
\usage{
\method{plot}{NDepositsPmf}(object, isMeanPlotted = TRUE, barwidth = 0.1,
  isUsgsStyle = TRUE)
}
\arguments{
\item{object}{An object of class "NDepositsPmf"}

\item{isMeanPlotted}{Logical variable indicating whether the mean is plotted on the pmf.}

\item{isUsgsStyle}{Make the plot format similar to, but not identical to, the
U.S. Geological Survey style}

\item{barWidth}{Width of the bars, which represent the probabilities for the numbers of
undiscovered deposits.}
}
\description{
Plots the probability mass function for the number of
undiscovered deposits in the permissive tract. If the type is Mark3 or Mark4,
then the alternative complementary cumulative distribution function
is included.
}
\details{
The alternative complementary cumulative distribution function is
G(x) = Pr( X >= x ) for - infinity < x < infinity. Between
the discrete values of X, the function G(x) is defined and would
be plotted as a horizontal lines. The problem is that these horizonal
lines will be confusing to people who are unfamilar with cumulative
distribution functions for discrete random variables. Consequently,
they are omitted.
}
\examples{
pmf1 <- NDepositsPmf("Poisson", list(theMean=5), "")
plot(pmf1)

pmf2 <- NDepositsPmf("Mark3", list(thresholds=c(1,7,20)), "")
plot(pmf2)

}
\references{
Grimmett, G., and Stirzaker, D., 2001, Probability and random processes,
3rd ed., Oxford University Press.
}

