% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TonnagePdf1.R
\name{plot.TonnagePdf1}
\alias{plot.TonnagePdf1}
\title{Plot the cdf for the material tonnages in a single,
undiscovered deposit}
\usage{
\method{plot}{TonnagePdf1}(object, whichMatTonnage = object$matNames,
  isUsgsStyle = TRUE)
}
\arguments{
\item{object}{An object of class "TonnagePdf1".}

\item{whichMatTonnage}{Character vector comprising the names of the materials that will be included
in the plot.}

\item{isUsgsStyle}{Make the plot format similar to, but not identical to, the
U.S. Geological Survey style}
}
\description{
Plot the cumulative distribution function (cdf)
for the material tonnages in a single, undiscovered deposit within the
permissive tract. Overlaid on the plot is the
empirical cumulative distribution function (ecdf)
for the observed material tonnages.
}
\details{
An internal call to function reshape2::melt generates the
message "No id variables; using all as measure variables", which should
be ignored.

In the plot, the solid line(s) represent the cdf, and the dots represent
the ecdf.
}
\examples{
pdf1 <- TonnagePdf1(ExampleTonnageData, "mt")
plot(pdf1)

}

