% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TonnagePdfPT.R
\name{plot.TonnagePdfPT}
\alias{plot.TonnagePdfPT}
\title{Plot the pdf for the material tonnages in the permissive tract}
\usage{
\method{plot}{TonnagePdfPT}(object, whichMatTonnage = object$matNames,
  isUsgsStyle = TRUE)
}
\arguments{
\item{object}{An object of class "TonnagePdf1".}

\item{whichMatTonnage}{Character vector comprising the names of the materials that will be included
in the plot.}

\item{isUsgsStyle}{Make the plot format similar to, but not identical to, the
U.S. Geological Survey style}
}
\description{
Plot the probability density function (pdf)
for the material tonnages in the
permissive tract.
}
\details{
An internal call to function reshape2::melt generates the
message "No id variables; using all as measure variables", which should
be ignored.
}
\examples{
pmf <- NDepositsPmf( "NegBinomial", list(theMean=5,theStdDev=4), "" )
pdf1 <- TonnagePdf1(ExampleTonnageData, "mt", pdfType = "normal")
pdfPT <- TonnagePdfPT(pmf, pdf1)
plot(pdfPT)

}

