% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Metadata.R
\name{print.Metadata}
\alias{print.Metadata}
\title{Print the meta data for the probability calculations}
\usage{
\method{print}{Metadata}(object)
}
\arguments{
\item{object}{An object of class "Metadata"}
}
\description{
Print the meta data for the probability calculations for one
permissive tract.
}
\examples{
metaData <- Metadata("Lucky Strike PT", "PT001")
print(metaData)

}

