% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NDepositsPmf.R
\name{summary.NDepositsPmf}
\alias{summary.NDepositsPmf}
\title{Summarize the pmf for the number of undiscovered deposits}
\usage{
\method{summary}{NDepositsPmf}(object)
}
\arguments{
\item{object}{An object of class "NDepositsPmf"}
}
\description{
Print a statistical summary of the probability mass function for
the number of undiscovered deposits in the permissive tract.
}
\examples{
pmf <- NDepositsPmf( "Poisson", list(theMean=5), "" )
summary( pmf )

}

