% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TonnagePdf1.R
\name{summary.TonnagePdf1}
\alias{summary.TonnagePdf1}
\title{Summarize the pdf for the material tonnages in a single, undiscovered
deposit}
\usage{
\method{summary}{TonnagePdf1}(object, nDigits = 2)
}
\arguments{
\item{object}{An object of class "TonnagePdf1"}

\item{nDigits}{Number of signficant digits.}
}
\description{
Summarize the probability density function (pdf) for the
material tonnages in a single, undiscovered deposit within the permissive
tract. Summary statistics are calculated for both
the observed material tonnages and
the pdf that represents those tonnages.
}
\details{
The summary statistics for the pdf are calculated from
random samples of that pdf.

It is common that the summary statistics for the observed material
tonnages differ somewhat from the summary statistics for the pdf, especially
when the pdf is truncated. The reason for the difference is that tonnages
typically have an enormous range, making the summary statistics somewhat
non-robust.
}
\examples{
pdf1 <- TonnagePdf1(ExampleTonnageData, "mt")
summary(pdf1)

}

