% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TonnagePdfPT.R
\name{summary.TonnagePdfPT}
\alias{summary.TonnagePdfPT}
\title{Summarize the pdf for the material tonnages in the permissive tract}
\usage{
\method{summary}{TonnagePdfPT}(object, nDigits = 2)
}
\arguments{
\item{object}{An object of class "TonnagePdfPT"}

\item{nDigits}{Number of signficant digits.}
}
\description{
Summarize the probability density function (pdf) for the
material tonnages in all undiscovered deposits within the permissive
tract.
}
\details{
The summary statistics for the pdf are calculated from
random samples of the pdf.
}
\examples{
pmf <- NDepositsPmf( "NegBinomial", list(theMean=5,theStdDev=4), "" )
pdf1 <- TonnagePdf1(ExampleTonnageData, "mt", pdfType = "normal")
pdfPT <- TonnagePdfPT(pmf, pdf1)
summary(pdfPT)

}

