% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TonnagePdf1.R
\name{TonnagePdf1}
\alias{TonnagePdf1}
\title{Construct the pdf for the material tonnages in a single,
undiscovered deposit}
\usage{
TonnagePdf1(obsTonnages, units, pdfType = "empirical", isTruncated = TRUE,
  minNDeposits = 20, nRandomSamples = 20000, seed = 7)
}
\arguments{
\item{obsTonnages}{Dataframe containing the observed material tonnages. (See details.)}

\item{units}{Character string containing units for the material tonnages.}

\item{pdfType}{Character string containing the type of pdf for
the log-transformed material tonnages. The choices are either
"empirical" or "normal".}

\item{isTruncated}{Logical variable indicating whether the pdf is
truncated at the lowest and highest values of the observed material tonnages.}

\item{minNDeposits}{Minimum number of deposits with observed material tonnages.}

\item{nRandomSamples}{Number of random samples used to compute summary statistics and the
cumulative distribution function.}

\item{seed}{Seed for the random number generator.}
}
\value{
If the input arguments have an error, the R-value NULL is returned.
Otherwise, a list with the following components is returned.

\item{obsTonnages}{Input argument obsTonnages.}

\item{units}{Input argument units}

\item{pdfType}{Input argument pdfType.}

\item{isTruncated}{Input argument isTruncated.}

\item{matNames}{Names of the materials.}

\item{logTonnages}{Matrix with the natural logarithm
transformation of the observed material tonnages.}

\item{theObsMean}{Mean vector for the observed tonnages.}

\item{theObsCov}{Covariance matrix for the observed tonnages.}

\item{theObsSd}{Standard deviation vector for the observed tonnages.}

\item{theObsCor}{Correlation matrix for the observed tonnages.}

\item{rs}{Random samples of the pdf.}

\item{theMean}{Mean vector for the pdf.}

\item{theCov}{Covariance matrix for the pdf.}

\item{theSd}{Standard deviation vector for the pdf.}

\item{theCor}{Correlation matrix for the pdf.}

\item{call}{Function call.}
}
\description{
Construct the probability density function (pdf) for the
material tonnages in a single, undiscovered deposit within the permissive
tract. The pdf is not explicitly specified; instead it is implicitly
specified with the random samples that are generated from it.
}
\details{
Dataframe obsTonnages comprises the material tonnages for the known
deposits. Each row comprise the data for one deposit. Each column comprises
a particular aspect of the data. The first column always lists the names
of the deposits. The name of the first column is always "Name".
The second column and subsequent columns, if any, list
the tonnages for each type of material. For example, if there is only
one material (which would correspond to only one resource), then its
tonnages would be listed in the second column. The name of this column
is the name of the resource. For example, it might be "Sand".
If there are two materials
(which might correspond to ore and a resource), then the ore tonnages would
be listed in the second column and the resource tonnages in the third
column. Again, the names of the second and third columns are the names of the
materials. For example, they might be "Ore" and "Copper".
If there are three or more materials, then the format of the
dataframe is similar.

All material tonnages in the dataframe must be greater than zero and
must not be missing.

The minimum number of
deposits (that is, rows in the dataframe) should be greater than or
equal to 20.
}
\examples{
pdf1 <- TonnagePdf1(ExampleTonnageData, "mt")

}

