% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TonnagePdfPT.R
\name{TonnagePdfPT}
\alias{TonnagePdfPT}
\title{Construct the pdf for the material tonnages in the
permissive tract}
\usage{
TonnagePdfPT(oPmf, oPdf, nRandomSamples = 20000)
}
\arguments{
\item{oPmf}{An object of class "NDepositsPmf"}

\item{oPdf}{An object of class "TonnagePdf1"}

\item{nRandomSamples}{Number of random samples of the material tonnages in the
permissive tract. The actual number may differ slightly because of
rounding.}
}
\value{
If the input arguments have an error, the R-value NULL is returned.
Otherwise, a list with the following components is returned.

\item{totalTonnages}{Matrix comprising random samples of
the total material tonnages in the permissive tract and the associated
number of undiscovered deposits.}

\item{units}{Input argument units}

\item{matNames}{Names of the materials.}

\item{theMean}{Mean vector for the total material tonnages in
the permissive tract.}

\item{theCov}{Covariance matrix for the total material tonnages in
the permissive tract.}

\item{theCor}{Correlation matrix for the total material tonnages in
the permissive tract.}

\item{theSd}{Standard deviation vector for the total material
tonnages in the permissive tract.}

\item{predMean}{Predicted mean vector for the total material
tonnages in the permissive tract.}

\item{predCov}{Predicted covariance matrix for the total material
tonnages in the permissive tract.}

\item{predCor}{Predicted correlation matrix for the total material
tonnages in the permissive tract.}

\item{predSd}{Predicted standard deviation vector for the total
material tonnages in the permissive tract.}

\item{call}{Function call}
}
\description{
Construct the probability density function (pdf) for the
material tonnages in the permissive
tract. The pdf is not explicitly specified; instead it is implicitly
specified with the random samples.
}
\details{
The random samples are computed using the algorithm in Ellefsen and others
(2016). The predicted mean vector and the predicted covariance matrix for
the pdf are computed using the analytic formulas in Ellefsen and others
(2016). The predicted correlation matrix and the predicted standard
deviation vector are calculated from the predicted covariance matrix.
}
\examples{
pmf <- NDepositsPmf( "NegBinomial", list(theMean=5,theStdDev=4), "" )
pdf1 <- TonnagePdf1(ExampleTonnageData, "mt", pdfType = "normal")
pdfPT <- TonnagePdfPT(pmf, pdf1)

}
\references{
Ellefsen, K.J., Phillips, J.D., Mihalasky, M.J., and Robinson, G.R., Jr.,
2016, Probability calculations for three-part mineral assessments,
U.S. Geological Survey Report XXXX
}

